+++ apps/api/src/app/app.module.ts
@@ -1,18 +1,19 @@
 import { Module } from '@nestjs/common';
 import { ConfigModule } from '@nestjs/config';
 import { configServiceFactory } from '../config/config.service';
-import { TermModule } from '../domain/term/term/term.module';
+import { DomainServicesModule } from '../domain/services/domain-services.module';
 import { AppController } from './app.controller';
 import { AppService } from './app.service';
+import { TermController } from './controllers/term.controller';
 
 @Module({
   imports: [
     ConfigModule.forRoot({
       load: [configServiceFactory],
     }),
-    TermModule,
+    DomainServicesModule,
   ],
-  controllers: [AppController],
+  controllers: [AppController, TermController],
   providers: [AppService],
 })
 export class AppModule {}
