+++ apps/api/src/app/controllers/term.controller.spec.ts
@@ -1,9 +1,14 @@
 import { Test, TestingModule } from '@nestjs/testing';
-import { DatabaseProvider } from 'apps/api/src/database/database.provider';
+import { TermService } from '../../domain/services/term.service';
+import { buildMockDatabaseProvicer } from '../../domain/services/term.service.spec';
+import { DatabaseProvider } from '../../persistence/database/database.provider';
 import { TermController } from './term.controller';
-import { TermService } from './term.service';
-import { buildMockDatabaseProvicer } from './term.service.spec';
 
+/**
+ * TODO We probably want a proper integration \ smoke 'contract' test with a
+ * test case for each of our endpoints to verify things are working at a high-level.
+ * E.g. verify that a valid request sends a 400.
+ */
 describe('TermController', () => {
   let controller: TermController;
