+++ apps/api/src/domain/models/entity.spec.ts
