+++ apps/api/src/domain/models/term/term.module.ts
@@ -1,0 +1,10 @@
+// import { Module } from '@nestjs/common';
+// import { DatabaseModule } from 'apps/api/src/database/database.module';
+// import { TermService } from './term.service';
+
+// @Module({
+//   imports: [DatabaseModule],
+//   controllers: [],
+//   providers: [TermService],
+// })
+// export class TermModule {}
