+++ apps/api/src/domain/services/domain-services.module.ts
@@ -1,0 +1,10 @@
+import { Module } from '@nestjs/common';
+import { DatabaseModule } from '../../persistence/database/database.module';
+import { TermService } from './term.service';
+
+@Module({
+  imports: [DatabaseModule],
+  controllers: [],
+  providers: [TermService],
+})
+export class DomainServicesModule {}
