+++ apps/api/src/domain/types/all-entities.ts
@@ -1,4 +1,4 @@
-import { CreateTermDto } from '../term/term/dto/create-term.dto';
+import { CreateTermDto } from '../models/term/dto/create-term.dto';
 import { CreateVocabularyListDto } from '../vocabulary-list/dto/create-vocabulary-list.dto';
 
 export type AllCreateEntityDtosUnion = CreateTermDto | CreateVocabularyListDto;
