+++ apps/api/src/persistence/database/arango-database.spec.ts
