+++ apps/api/src/persistence/database/interfaces/database-provider.ts
