+++ apps/api/src/persistence/repositories/book-repository.ts
