+++ apps/api/src/view-models/term-view-model.ts
@@ -1,4 +1,4 @@
-import { Term } from '../domain/term/term/entities/term.entity';
+import { Term } from '../domain/models/term/entities/term.entity';
 
 export class TermViewModel {
   // TODO Map `Term` domain model to `Term` view model.
